backend/requirements.txt
# Core FastAPI & Web Framework
fastapi==0.115.3
uvicorn==0.40.0
starlette==0.41.0
pydantic==2.9.1
pydantic_core==2.23.3

# HTTP & Networking
requests==2.32.3
httpx==0.27.2
httpcore==1.0.5
aiohttp==3.10.5
urllib3==2.6.3

# Data Processing & Analysis
pandas==2.2.3
numpy==1.23.5
scikit-learn==1.2.2

# Solar/PV Calculations
pvlib==0.11.1

# Machine Learning & Model Loading
joblib==1.3.2
xgboost==1.7.5

# AI/LLM
google-generativeai==0.8.0
langdetect==1.0.9

# Database & Caching
redis==5.0.1

# Utilities
python-dotenv==1.0.1
pytz==2024.2
python-dateutil==2.9.0.post0

# Type Hints
typing_extensions==4.9.0
